backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
qdrant-client==1.12.0
asyncpg==0.30.0
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.52.0
tiktoken==0.8.0
sentence-transformers==3.1.1
PyYAML==6.0.2
python-multipart==0.0.12
beautifulsoup4==4.12.3
markdown==3.6
cohere==5.5.3